+++ src/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.2.3 (2017-11-16)
+
+** Ehancements:**
+
+- Add new template styles for "What's New"
+- New documentation website styling
+- Backend refactoring
+
 ## 0.2.2 (2017-11-08)
 
 **Enhancements:**